global/changed architecture of modules
