+++ src/app/app-routing.module.ts
@@ -1,9 +1,11 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { AuthComponent } from './components/auth/auth.component';
 import { WebsiteComponent } from './views/website/website.component';
 
 const routes: Routes = [
   { path: '', component: WebsiteComponent },
+  { path: 'auth', component: AuthComponent },
 ];
 
 @NgModule({
