+++ src/app/app.component.html
@@ -1,1 +1,2 @@
+<app-header></app-header>
 <router-outlet></router-outlet>
