+++ src/app/modules/app.module.ts
@@ -1,0 +1,26 @@
+import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+
+import { AppRoutingModule } from '../app-routing.module';
+import { AppComponent } from '../app.component';
+import { MentorPanelComponent } from '../components/mentor-panel/mentor-panel.component';
+import { HeaderComponent } from '../utils/UI/header/header.component';
+import { MentorPanelViewsComponent } from '../views/mentor-panel-views/mentor-panel-views.component';
+import { AuthModule } from './auth/auth.module';
+import { WebsiteModule } from './website/website.module';
+
+@NgModule({
+  declarations: [
+    AppComponent,
+    HeaderComponent,
+  ],
+  imports: [
+    AuthModule,
+    WebsiteModule,
+    BrowserModule,
+    AppRoutingModule
+  ],
+  providers: [],
+  bootstrap: [AppComponent]
+})
+export class AppModule { }
