+++ src/app/modules/auth/auth.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AuthComponent } from 'src/app/components/auth/auth.component';
+
+@NgModule({
+  declarations: [
+    AuthComponent,
+  ],
+  imports: [
+    CommonModule
+  ]
+})
+export class AuthModule { }
