+++ src/app/modules/website/website.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { WebsiteComponent } from 'src/app/views/website/website.component';
+
+@NgModule({
+  declarations: [
+    WebsiteComponent,
+  ],
+  imports: [
+    CommonModule
+  ]
+})
+export class WebsiteModule { }
