+++ src/app/utils/UI/header/header.component.html
@@ -1,1 +1,11 @@
-<p>header works!</p>
+<header>
+  <a routerLink="">
+    <img src="#" alt="logo">
+  </a>
+  <menu>
+    <ul>
+      <li> <a routerLink="/help">Допомога</a></li>
+      <li> <a routerLink="/auth">Вхід/Реєстрація</a></li>
+    </ul>
+  </menu>
+</header>
