+++ src/app/views/mentor-panel-views/mentor-panel-views.component.html
@@ -1,0 +1,1 @@
+<p>mentor-panel-views works!</p>
