+++ src/app/views/mentor-panel-views/mentor-panel-views.component.spec.ts
@@ -1,20 +1,20 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { HeaderComponent } from './header.component';
+import { MentorPanelViewsComponent } from './mentor-panel-views.component';
 
-describe('HeaderComponent', () => {
-  let component: HeaderComponent;
-  let fixture: ComponentFixture<HeaderComponent>;
+describe('MentorPanelViewsComponent', () => {
+  let component: MentorPanelViewsComponent;
+  let fixture: ComponentFixture<MentorPanelViewsComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ HeaderComponent ]
+      declarations: [ MentorPanelViewsComponent ]
     })
     .compileComponents();
   });
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(HeaderComponent);
+    fixture = TestBed.createComponent(MentorPanelViewsComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
