+++ src/app/views/mentor-panel-views/mentor-panel-views.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-mentor-panel-views',
+  templateUrl: './mentor-panel-views.component.html',
+  styleUrls: ['./mentor-panel-views.component.sass']
+})
+export class MentorPanelViewsComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
